+++ website/app/views/makepred.py
@@ -25,42 +25,48 @@ return: status, msg
 msg = filename if success
 '''
 def prepare_request(request):
-    # There is no input file in the request
-    if 'input-file' not in request.files:
-        return 'error','no input file part'
-    file = request.files['input-file']
-    # Check file size
-    file.seek(0, os.SEEK_END)
-    file_length = file.tell()
-    if file_length > app.config['MAX_FILE_LENGTH']:
-        maxsize = app.config['MAX_FILE_LENGTH'] / (1024*1024)
-        return 'error','uploaded file is larger than the allowed maximum size of %dMB' % maxsize
-    file.seek(0) # seek back
-    # Can only accept tsv or csv
-    if not allowed_file(file.filename):
-        return 'error','please upload only tsv or csv'
-    # No file selected:
-    if file.filename == '':
-        return 'error','no selected file'
-    # No TFs selected
-    if not request.form.getlist('pred-select'):
-        return 'error','please select transcription factors'
-    # Check if we have all the columns we need
-    filename = secure_filename(file.filename)
-    filepath = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-    file.save(filepath)
-    required = ['chromosome','chromosome_start','mutation_type','mutated_from_allele','mutated_to_allele']
-    with open(filepath) as f:
-        file_extension = os.path.splitext(filepath)[1]
-        if file_extension == ".tsv":
-            incols = f.readline().strip().split("\t")
-        else: # must be csv since we checked it
-            incols = f.readline().strip().split(",")
-    if not all(elem in incols for elem in required):
-        os.remove(filepath)
-        return 'error','some required fields are missing from the input file'
-    # Finally, file is okay
-    return 'success',filename
+    if request.form.get('input-mode') == "2":
+        if request.form.get('input-example-list') == "None":
+            return 'error','no input file part'
+        else:
+            return 'example', request.form.get('input-example-list')
+    else:
+        # There is no input file in the request
+        if 'input-file' not in request.files:
+            return 'error','no input file part'
+        file = request.files['input-file']
+        # Check file size
+        file.seek(0, os.SEEK_END)
+        file_length = file.tell()
+        if file_length > app.config['MAX_FILE_LENGTH']:
+            maxsize = app.config['MAX_FILE_LENGTH'] / (1024*1024)
+            return 'error','uploaded file is larger than the allowed maximum size of %dMB' % maxsize
+        file.seek(0) # seek back
+        # Can only accept tsv or csv
+        if not allowed_file(file.filename):
+            return 'error','please upload only tsv or csv'
+        # No file selected:
+        if file.filename == '':
+            return 'error','no selected file'
+        # No TFs selected
+        if not request.form.getlist('pred-select'):
+            return 'error','please select transcription factors'
+        # Check if we have all the columns we need
+        filename = secure_filename(file.filename)
+        filepath = os.path.join(app.config['UPLOAD_FOLDER'], filename)
+        file.save(filepath)
+        required = ['chromosome','chromosome_start','mutation_type','mutated_from_allele','mutated_to_allele']
+        with open(filepath) as f:
+            file_extension = os.path.splitext(filepath)[1]
+            if file_extension == ".tsv":
+                incols = f.readline().strip().split("\t")
+            else: # must be csv since we checked it
+                incols = f.readline().strip().split(",")
+        if not all(elem in incols for elem in required):
+            os.remove(filepath)
+            return 'error','some required fields are missing from the input file'
+        # Finally, file is okay
+        return 'success',filename
 
 
 @app.route('/upload', methods=['POST'])
@@ -72,7 +78,13 @@ def handle_upload():
         if status=='error':
             return jsonify({'Message':msg}), 500
         else:
+            if status == "example":
+                filepath = app.config['STATIC_EXAMPLE_DIR'] + msg
+            else:
+                filepath = app.config['UPLOAD_FOLDER'] + msg
             # request.form.getlist('pred-select'):['Arid3a:Arid3a_3875.1_v1_deBruijn', 'Bhlhb2:Bhlhb2_4971.1_v1_deBruijn']
+            print(request.form.getlist('pred-select'));
+            print(request.form.get('pred-select'));
             genes_selected = [elm.split(":")[0] for elm in request.form.getlist('pred-select')]
 
             select_list = [elm.split(":")[1] for elm in request.form.getlist('pred-select')]
@@ -86,7 +98,7 @@ def handle_upload():
             else:
                 filterval = float(request.form.get('output-selection-opt'))
 
-            task = chain(celerytask.inittbl.s(app.config['UPLOAD_FOLDER'] + msg,
+            task = chain(celerytask.inittbl.s(filepath,
                         app.config['CHRDIR'] +"/"+chrver),
                         celerytask.do_prediction.s(unique_pbms,genes_selected,filteropt,filterval)).apply_async() # put genes_selected here
 
@@ -132,10 +144,8 @@ def get_examplelist():
 
 #==================
 
-@app.route('/makepred', methods=['GET', 'POST'])
+@app.route('/makeprediction', methods=['GET', 'POST'])
 def makepred():
-    #session.permanent = True
-    #session.clear() -- need to limit the amount of session somewhere
     return render_template("makepred.html")
 
 @app.route('/download/<path:filename>')
